OOP for CrossingGuardBot
